design: 화면 디자인에 맞게 컬러 변경 #68
